# swagger.yaml
openapi: 3.0.0
info:
  title: Flow Tester API
  version: 1.0.0
  description: API for managing flow testing applications and environments
servers:
  - url: http://localhost:3000
    description: Local development server
paths:
...
  /applications/{application}:
    get:
      summary: Get specific application
      parameters:
        - name: application
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Application details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Application'
  /applications/{application}/envs:
    get:
      summary: Get application environments
      parameters:
        - name: application
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of environment files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EnvFile'
  /applications/{application}/envs/{env}:
    get:
      summary: Get specific environment file
      parameters:
        - name: application
          in: path
          required: true
          schema:
            type: string
        - name: env
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Environment file details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvFile'
...
components:
  schemas:
    Application:
      type: object
      properties:
        slug:
          type: string
        envFiles:
          type: array
          items:
            $ref: '#/components/schemas/EnvFile'
    EnvFile:
      type: object
      properties:
        name:
          type: string
        path:
          type: string